chore: created a separate component for navlinks
