+++ src/admin/components/elements/Header/types.ts
@@ -1,0 +1,5 @@
+import { Collections } from '@/admin/data/collections/types'
+
+export type Props = {
+  navLinks?: Collections
+}
