+++ src/admin/components/elements/Nav/index.tsx
@@ -2,10 +2,9 @@
 
 import { BellIcon, Package2Icon } from '../../icons'
 import Link from 'next/link'
-import { usePathname } from 'next/navigation'
 
 import { collections } from '@/admin/data/collections'
-import { CollectionItem, Collections } from '@/admin/data/collections/types'
+import { Collections } from '@/admin/data/collections/types'
 import { Button } from '@/components/ui/button'
 import {
   Card,
@@ -15,10 +14,10 @@ import {
   CardTitle,
 } from '@/components/ui/card'
 
+import NavLinks from './NavLinks'
 import { Props } from './types'
 
 const DefaultNav: React.FC<Props> = props => {
-  const pathname = usePathname()
   const { navLinks } = props
 
   return (
@@ -28,7 +27,7 @@ const DefaultNav: React.FC<Props> = props => {
         <div className='flex h-14 items-center border-b px-4 lg:h-[60px] lg:px-6'>
           <Link className='flex items-center gap-2 font-semibold' href='#'>
             <Package2Icon className='h-6 w-6' />
-            <span className=''>Acme Inc</span>
+            <span className=''>Dev Cms</span>
           </Link>
           <Button className='ml-auto h-8 w-8' size='icon' variant='outline'>
             <BellIcon className='h-4 w-4' />
@@ -36,27 +35,7 @@ const DefaultNav: React.FC<Props> = props => {
           </Button>
         </div>
         <div className='flex-1'>
-          <nav className='grid items-start px-2 text-sm font-medium lg:px-4'>
-            {!navLinks?.length && (
-              <h3 className='place-self-center pt-5'>No Collections</h3>
-            )}
-            {navLinks?.map(({ slug, label, url, icon }: CollectionItem) => {
-              const isActive = pathname === url
-              const linkClassNames = isActive
-                ? 'flex items-center gap-3 rounded-lg bg-muted px-3 py-2 text-primary transition-all hover:text-primary'
-                : 'flex items-center gap-3 rounded-lg px-3 py-2 text-muted-foreground transition-all hover:text-primary'
-
-              return (
-                <Link
-                  key={url}
-                  className={linkClassNames}
-                  href={url || `/admin/collections/${slug}`}>
-                  {icon}
-                  {label}
-                </Link>
-              )
-            })}
-          </nav>
+          <NavLinks navLinks={navLinks} />
         </div>
         <div className='mt-auto p-4'>
           <Card x-chunk='dashboard-02-chunk-0'>
